#8
月売目標と、売上を表示対応完了

+++ database/migrations/2025_01_05_061830_create_store_sales_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('store_id')->comment('店舗ID')->constrained('stores')->onDelete('cascade');
             $table->date('date')->comment('月売の日付');
-            $table->decimal('sales_amount', 15, 2)->comment('売目金額');
+            $table->integer('sales_amount')->comment('売目金額');
             $table->timestamps();
         });
     }
